+++ packages/contracts/src/http/request.ts
@@ -1,5 +1,6 @@
 'use strict'
 
+import { IncomingHttpHeaders } from 'http'
 import { InteractsWithContentTypes } from './concerns'
 
 export interface HttpRequest extends InteractsWithContentTypes {
@@ -31,7 +32,7 @@ export interface HttpRequest extends InteractsWithContentTypes {
   /**
    * Returns the request headers.
    */
-  headers: any
+  headers: IncomingHttpHeaders
 
   /**
    * Returns the request header identified by the given `key`. The default